Convert AdditionTile to TypeScript
